+++ scripts/BaseGrid.js
@@ -16,6 +16,10 @@ PATCHES.BASIC = {};
  */
 function _getRulerDestination(wrapped, ray, offset, token) {
   if ( canvas.controls.ruler._unsnap ) return ray.B.add(offset);
+
+  // We are moving from the token center, so add back 1/2 width/height to offset.
+  offset.x += canvas.scene.dimensions.size * 0.5;
+  offset.y += canvas.scene.dimensions.size * 0.5;
   return wrapped(ray, offset, token);
 }